\name{rdbb-package}
\alias{rdbb-package}
\alias{rdbb}
\docType{package}
\title{
\packageTitle{rdbb}
}
\description{
\packageDescription{rdbb}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{rdbb}

Maintainer: \packageMaintainer{rdbb}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rdbb)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rdbb}}
%\section{Documentation Index}{\packageIndices{rdbb}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
